route tab as a subpath, not a search param
